requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
repoze.lru==0.7
requests==2.24.0
Routes==2.4.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
serializers==0.2.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
